+++ python-flask/app/models/post.py
@@ -1,0 +1,9 @@
+from app.extensions import db
+
+class Message(db.Model):
+    __tablename__ = 'messages'
+    message_id = db.Column(db.Integer, primary_key=True)
+    author_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
+    text = db.Column(db.Text, nullable=False)
+    pub_date = db.Column(db.Text)
+    flagged = db.Column(db.Integer)
