+++ python-flask/app/templates/layout.html
@@ -5,13 +5,13 @@
   <h1>MiniTwit</h1>
   <div class=navigation>
   {% if g.user %}
-    <a href="{{ url_for('timeline') }}">my timeline</a> |
-    <a href="{{ url_for('public_timeline') }}">public timeline</a> |
-    <a href="{{ url_for('logout') }}">sign out [{{ g.user.username }}]</a>
+    <a href="{{ url_for('main.timeline') }}">my timeline</a> |
+    <a href="{{ url_for('main.public_timeline') }}">public timeline</a> |
+    <a href="{{ url_for('admin.logout') }}">sign out [{{ g.user.username }}]</a>
   {% else %}
-    <a href="{{ url_for('public_timeline') }}">public timeline</a> |
-    <a href="{{ url_for('register') }}">sign up</a> |
-    <a href="{{ url_for('login') }}">sign in</a>
+    <a href="{{ url_for('main.public_timeline') }}">public timeline</a> |
+    <a href="{{ url_for('admin.register') }}">sign up</a> |
+    <a href="{{ url_for('admin.login') }}">sign in</a>
   {% endif %}
   </div>
   {% with flashes = get_flashed_messages() %}
