+++ python-flask/wsgi.py
@@ -1,4 +1,6 @@
-from minitwit import app
+from app import create_app
 
-if __name__ == "__main__":
-    app.run()
+app = create_app()
+
+if __name__ == '__main__':
+    app.run(host="0.0.0.0", port=5000)
